@@ -1,8 +1,8 @@
-import os
 import logging
-from typing import Mapping
 
-from organize.compat import Path
+from fs.base import FS
+
+from organize.utils import safe_description
 
 from .action import Action
 
@@ -16,29 +16,22 @@ class Delete(Action):
 
     Deleted files have no recovery option!
     Using the `Trash` action is strongly advised for most use-cases!
-
-    Example:
-        - Delete all JPGs and PNGs on the desktop which are older than one year:
-
-          .. code-block:: yaml
-            :caption: config.yaml
-
-            rules:
-              - folders: '~/Desktop'
-              - filters:
-                  - lastmodified:
-                      - days: 365
-                  - extension:
-                      - png
-                      - jpg
-              - actions:
-                  - delete
     """
 
-    def pipeline(self, args: Mapping):
-        path = args["path"]  # type: Path
-        simulate = args["simulate"]  # type: bool
-        self.print('Delete "%s"' % path)
+    name = "delete"
+
+    @classmethod
+    def get_schema(cls):
+        return cls.name
+
+    def pipeline(self, args: dict, simulate: bool):
+        fs = args["fs"]  # type: FS
+        fs_path = args["fs_path"]  # type: str
+        desc = safe_description(fs=fs, path=fs_path)
+        self.print(f'Deleting "{desc}"')
         if not simulate:
-            logger.info("Deleting file %s.", path)
-            os.remove(str(path))
+            logger.info("Deleting %s.", desc)
+            if fs.isdir(fs_path):
+                fs.removetree(fs_path)
+            elif fs.isfile(fs_path):
+                fs.remove(fs_path)
