@@ -1,7 +1,4 @@
 import logging
-from typing import Mapping
-
-from organize.compat import Path
 
 from .action import Action
 
@@ -10,34 +7,23 @@ logger = logging.getLogger(__name__)
 
 class Trash(Action):
 
-    """
-    Move a file into the trash.
+    """Move a file or dir into the trash."""
 
-    Example:
-        - Move all JPGs and PNGs on the desktop which are older than one year
-          into the trash:
+    name = "trash"
 
-          .. code-block:: yaml
-            :caption: config.yaml
+    @classmethod
+    def get_schema(cls):
+        return cls.name
 
-            rules:
-              - folders: '~/Desktop'
-              - filters:
-                  - lastmodified:
-                      - days: 365
-                  - extension:
-                      - png
-                      - jpg
-              - actions:
-                  - trash
-    """
+    def trash(self, path: str, simulate: bool):
+        from send2trash import send2trash
 
-    def pipeline(self, args: Mapping):
-        path = args["path"]  # type: Path
-        simulate = args["simulate"]  # type: bool
-        from send2trash import send2trash  # type: ignore
-
-        self.print('Trash "%s"' % path)
+        self.print(f'Trash "{path}"')
         if not simulate:
             logger.info("Moving file %s into trash.", path)
-            send2trash(str(path))
+            send2trash(path)
+
+    def pipeline(self, args: dict, simulate: bool):
+        fs = args["fs"]
+        fs_path = args["fs_path"]
+        self.trash(path=fs.getsyspath(fs_path), simulate=simulate)
