@@ -1,4 +1,4 @@
-from organize.cli import main
+from organize.cli import cli
 
 if __name__ == "__main__":
-    main()
+    cli()
