@@ -1,38 +1,14 @@
 import logging
 import logging.config
-import os
 
-import appdirs  # type: ignore
-import colorama  # type: ignore
 import yaml
+from fs import appfs
 
-from .compat import Path
-
-colorama.init(autoreset=True)
-
-# prepare config and log folders
-APP_DIRS = appdirs.AppDirs("organize")
-
-# setting the $ORGANIZE_CONFIG env variable overrides the default config path
-if os.getenv("ORGANIZE_CONFIG"):
-    CONFIG_PATH = Path(os.getenv("ORGANIZE_CONFIG", "")).resolve()
-    CONFIG_DIR = CONFIG_PATH.parent
-else:
-    CONFIG_DIR = Path(APP_DIRS.user_config_dir)
-    CONFIG_PATH = CONFIG_DIR / "config.yaml"
-
-LOG_DIR = Path(APP_DIRS.user_log_dir)
-LOG_PATH = LOG_DIR / "organize.log"
-
-for folder in (CONFIG_DIR, LOG_DIR):
-    folder.mkdir(parents=True, exist_ok=True)
-
-# create empty config file if it does not exist
-if not CONFIG_PATH.exists():
-    CONFIG_PATH.touch()
+with appfs.UserLogFS("organize") as log_fs:
+    LOG_PATH = log_fs.getsyspath("organize.log")
 
 # configure logging
-LOGGING_CONFIG = """
+LOGGING_CONFIG = f"""
 version: 1
 disable_existing_loggers: false
 formatters:
@@ -47,7 +23,8 @@ handlers:
     file:
         class: logging.handlers.TimedRotatingFileHandler
         level: DEBUG
-        filename: {filename}
+        filename: {str(LOG_PATH)}
+        encoding: utf-8
         formatter: simple
         when: midnight
         backupCount: 30
@@ -56,7 +33,5 @@ root:
     handlers: [file]
 exifread:
     level: INFO
-""".format(
-    filename=str(LOG_PATH)
-)
+"""
 logging.config.dictConfig(yaml.safe_load(LOGGING_CONFIG))
