@@ -1,6 +1,4 @@
-import sys
-
 if __name__ == "__main__":
-    from .cli import main
+    from .cli import cli
 
-    sys.exit(main())
+    cli()
