@@ -1,8 +1,9 @@
 import logging
 import textwrap
-from typing import Any, Mapping, Optional, Iterable
+from typing import Any, Dict, Iterable
+from typing import Optional as tyOptional
 
-from organize.utils import DotDict
+from schema import Optional, Or
 
 from .action import Action
 
@@ -11,62 +12,28 @@ logger = logging.getLogger(__name__)
 
 class Python(Action):
 
-    r"""
-    Execute python code in your config file.
+    """Execute python code.
 
-    :param str code: The python code to execute
-
-    Examples:
-      - A basic example that shows how to get the current file path and do some
-        printing in a for loop. The ``|`` is yaml syntax for defining a string
-        literal spanning multiple lines.
-
-        .. code-block:: yaml
-            :caption: config.yaml
-
-            rules:
-            - folders: '~/Desktop'
-              actions:
-                - python: |
-                    print('The path of the current file is %s' % path)
-                    for _ in range(5):
-                        print('Heyho, its me from the loop')
-
-      - You can access filter data:
-
-        .. code-block:: yaml
-            :caption: config.yaml
-
-            rules:
-            - folders: ~/Desktop
-              filters:
-                - regex: '^(?P<name>.*)\.(?P<extension>.*)$'
-              actions:
-                - python: |
-                    print('Name: %s' % regex.name)
-                    print('Extension: %s' % regex.extension)
-
-      - You have access to all the python magic -- do a google search for each
-        filename starting with an underscore:
-
-        .. code-block:: yaml
-            :caption: config.yaml
-
-            rules:
-            - folders: ~/Desktop
-              filters:
-                - filename:
-                    startswith: '_'
-              actions:
-                - python: |
-                    import webbrowser
-                    webbrowser.open('https://www.google.com/search?q=%s' % path.stem)
+    Args:
+        code (str): The python code to execute.
+        run_in_simulation (bool):
+            Whether to execute this code in simulation mode (Default false).
     """
 
-    def __init__(self, code) -> None:
-        self.code = textwrap.dedent(code)
+    name = "python"
+    arg_schema = Or(
+        str,
+        {
+            "code": str,
+            Optional("run_in_simulation"): bool,
+        },
+    )
 
-    def usercode(self, *args, **kwargs) -> Optional[Any]:
+    def __init__(self, code, run_in_simulation=False) -> None:
+        self.code = textwrap.dedent(code)
+        self.run_in_simulation = run_in_simulation
+
+    def usercode(self, *args, **kwargs):
         pass  # will be overwritten by `create_method`
 
     def create_method(self, name: str, argnames: Iterable[str], code: str) -> None:
@@ -80,13 +47,12 @@ class Python(Action):
         )
         exec(funccode, globals_, locals_)  # pylint: disable=exec-used
 
-    def pipeline(self, args: DotDict) -> Optional[Mapping[str, Any]]:
-        simulate = args.simulate
-        if simulate:
-            self.print("Code not run in simulation. (Args: %s)" % args)
+    def pipeline(self, args: dict, simulate: bool) -> tyOptional[Dict[str, Any]]:
+        if simulate and not self.run_in_simulation:
+            self.print("** Code not run in simulation. **")
             return None
 
-        logger.info('Executing python:\n"""\n%s\n""", args=%s', self.code, args)
+        logger.info('Executing python:\n"""\n%s\n"""', self.code)
         self.create_method(name="usercode", argnames=args.keys(), code=self.code)
         self.print("Running python script.")
 
