@@ -1,116 +1,59 @@
-import sys
-from datetime import datetime, timedelta
-from typing import Dict, Optional, SupportsFloat
+import subprocess
+from datetime import datetime, timezone
 
-from organize.compat import Path
-from organize.utils import DotDict
+from fs.base import FS
 
-from .filter import Filter
+from ._timefilter import TimeFilter
 
 
-class Created(Filter):
+class Created(TimeFilter):
+    """Matches files / folders by created date
 
-    """
-    Matches files by created date
+    Args:
+        years (int): specify number of years
+        months (int): specify number of months
+        weeks (float): specify number of weeks
+        days (float): specify number of days
+        hours (float): specify number of hours
+        minutes (float): specify number of minutes
+        seconds (float): specify number of seconds
+        mode (str):
+            either 'older' or 'newer'. 'older' matches files / folders created before the given
+            time, 'newer' matches files / folders created within the given time.
+            (default = 'older')
 
-    :param int days:
-        specify number of days
-
-    :param int hours:
-        specify number of hours
-
-    :param int minutes:
-        specify number of minutes
-
-    :param str mode:
-        either 'older' or 'newer'. 'older' matches all files created before the given
-        time, 'newer' matches all files created within the given time.
-        (default = 'older')
-
-    :returns:
-        - ``{created.year}`` -- the year the file was created
-        - ``{created.month}`` -- the month the file was created
-        - ``{created.day}`` -- the day the file was created
-        - ``{created.hour}`` -- the hour the file was created
-        - ``{created.minute}`` -- the minute the file was created
-        - ``{created.second}`` -- the second the file was created
-
-    Examples:
-        - Show all files on your desktop created at least 10 days ago:
-
-          .. code-block:: yaml
-            :caption: config.yaml
-
-            rules:
-              - folders: '~/Desktop'
-                filters:
-                  - created:
-                      days: 10
-                actions:
-                  - echo: 'Was created at least 10 days ago'
-
-        - Show all files on your desktop which were created within the last 5 hours:
-
-          .. code-block:: yaml
-            :caption: config.yaml
-
-            rules:
-              - folders: '~/Desktop'
-                filters:
-                  - created:
-                      hours: 5
-                      mode: newer
-                actions:
-                  - echo: 'Was created within the last 5 hours'
-
-        - Sort pdfs by year of creation:
-
-          .. code-block:: yaml
-            :caption: config.yaml
-
-            rules:
-              - folders: '~/Documents'
-                filters:
-                  - extension: pdf
-                  - created
-                actions:
-                  - move: '~/Documents/PDF/{created.year}/'
+    Returns:
+        {created}: The datetime the file / folder was created.
     """
 
-    def __init__(self, days=0, hours=0, minutes=0, seconds=0, mode="older") -> None:
-        self._mode = mode.strip().lower()
-        if self._mode not in ("older", "newer"):
-            raise ValueError("Unknown option for 'mode': must be 'older' or 'newer'.")
-        else:
-            self.is_older = self._mode == "older"
-        self.timedelta = timedelta(
-            days=days, hours=hours, minutes=minutes, seconds=seconds
-        )
+    name = "created"
 
-    def pipeline(self, args: DotDict) -> Optional[Dict[str, datetime]]:
-        created_date = self._created(args.path)
-        reference_date = datetime.now() - self.timedelta
-        match = (self.is_older and created_date <= reference_date) or (
-            not self.is_older and created_date >= reference_date
-        )
-        if match:
-            return {"created": created_date}
-        return None
+    def get_datetime(self, args) -> datetime:
+        fs = args["fs"]  # type: FS
+        fs_path = args["fs_path"]
+        created = fs.getinfo(fs_path, namespaces=["details"]).created
+        if not created:
+            # We're probably on Linux. No easy way to get creation dates here,
+            # so we try to use the stat utility.
+            created = self.fallback_method(fs, fs_path)
+        if not created:
+            raise EnvironmentError("File creation time is not available.")
+        return created
 
-    def _created(self, path: Path) -> datetime:
-        # see https://stackoverflow.com/a/39501288/300783
-        stat = path.stat()
-        time = 0  # type: SupportsFloat
-        if sys.platform.startswith("win"):
-            time = stat.st_ctime
-        else:
-            try:
-                time = stat.st_birthtime
-            except AttributeError:
-                # We're probably on Linux. No easy way to get creation dates here,
-                # so we'll settle for when its content was last modified.
-                time = stat.st_mtime
-        return datetime.fromtimestamp(float(time))
+    def fallback_method(self, fs, fs_path):
+        if fs.hassyspath(fs_path):
+            syspath = fs.getsyspath(fs_path)
+            commands = (
+                ["stat", "--format=%W", syspath],  # GNU coreutils
+                ["stat", "-f %B", syspath],  # BSD
+            )
+            for cmd in commands:
+                try:
+                    created_str = subprocess.check_output(cmd, encoding="utf-8").strip()
+                    timestamp = int(created_str)
+                    return datetime.fromtimestamp(timestamp, tz=timezone.utc)
+                except subprocess.CalledProcessError:
+                    pass
 
     def __str__(self):
-        return "Created(delta=%s, select_mode=%s)" % (self.timedelta, self._mode)
+        return "<Created>"
