@@ -1,27 +1,208 @@
+import logging
 import os
-import re
-from collections.abc import Mapping
 from copy import deepcopy
-from typing import Any, Sequence, Tuple, Union, List, Hashable
+from datetime import datetime
+from typing import Any, List, Sequence, Tuple, Union
 
-from .compat import Path
-
-WILDCARD_REGEX = re.compile(r"(?<!\\)[\*\?\[]+")
-
-
-def splitglob(globstr: str) -> Tuple[Path, str]:
-    """ split a string with wildcards into a base folder and globstring """
-    path = fullpath(globstr.strip())
-    parts = path.parts
-    for i, part in enumerate(parts):
-        if WILDCARD_REGEX.search(part):
-            return (Path(*parts[:i]), str(Path(*parts[i:])))
-    return (path, "")
+import jinja2
+from fs import open_fs
+from fs import path as fspath
+from fs.base import FS
+from fs.memoryfs import MemoryFS
+from jinja2 import nativetypes
 
 
-def fullpath(path: Union[str, Path]) -> Path:
-    """ Expand '~' and resolve the given path. Path can be a string or a Path obj. """
-    return Path(os.path.expandvars(str(path))).expanduser().resolve(strict=False)
+def finalize_placeholder(x):
+    # This is used to make the `path` arg available in the filters and actions.
+    # If a template uses `path` where no syspath is available this makes it possible
+    # to raise an exception.
+    if isinstance(x, Exception):
+        raise x
+    return x
+
+
+Template = jinja2.Environment(
+    variable_start_string="{",
+    variable_end_string="}",
+    autoescape=False,
+    finalize=finalize_placeholder,
+)
+
+NativeTemplate = nativetypes.NativeEnvironment(
+    variable_start_string="{",
+    variable_end_string="}",
+    autoescape=False,
+    finalize=finalize_placeholder,
+)
+
+
+def basic_args():
+    """The basic args which are guaranteed to be available."""
+    return {
+        "env": os.environ,
+        "now": datetime.now(),
+        "utcnow": datetime.utcnow(),
+    }
+
+
+class SimulationFS(MemoryFS):
+    def __init__(self, fs_url, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.fs_url = fs_url
+
+    def __str__(self):
+        if not self.fs_url:
+            return "<Simulation>"
+        elif "://" in self.fs_url:
+            return f"<{self.fs_url}>"
+        return self.fs_url
+
+
+def open_fs_or_sim(fs_url, *args, simulate=False, **kwargs):
+    if simulate:
+        simFS = SimulationFS(fs_url)
+        return simFS
+    return open_fs(fs_url, *args, **kwargs)
+
+
+def expand_user(fs_url: str) -> str:
+    fs_url = os.path.expanduser(fs_url)
+    if fs_url.startswith("zip://~"):
+        fs_url = fs_url.replace("zip://~", "zip://" + os.path.expanduser("~"))
+    elif fs_url.startswith("tar://~"):
+        fs_url = fs_url.replace("tar://~", "tar://" + os.path.expanduser("~"))
+    return fs_url
+
+
+def expand_args(template: Union[str, jinja2.environment.Template], args=None):
+    if not args:
+        args = basic_args()
+
+    if isinstance(template, str):
+        text = Template.from_string(template).render(**args)
+    else:
+        text = template.render(**args)
+
+    # expand user and fill environment vars
+    text = expand_user(text)
+    text = os.path.expandvars(text)
+
+    return text
+
+
+def fs_path_from_options(
+    path: str, filesystem: Union[FS, str, None] = ""
+) -> Tuple[FS, str]:
+    """
+    path can be a fs_url a normal fs_path
+    filesystem is optional and may be a fs_url.
+
+    - user tilde is expanded
+    - if a filesystem is given, we use that.
+    - otherwise we treat the path as a filesystem.
+    """
+    path = expand_args(path)
+
+    if filesystem:
+        if isinstance(filesystem, str):
+            filesystem = expand_args(filesystem)
+            return (open_fs(filesystem), path)  # type: ignore
+        return (filesystem, path)
+    return (open_fs(path), "/")
+
+
+def unwrap_wrapfs(fs, path):
+    from fs.path import abspath
+    from fs.wrapfs import WrapFS
+
+    if isinstance(fs, WrapFS):
+        fs, path = fs.delegate_path(path)
+    return fs, abspath(path)
+
+
+def is_same_resource(fs1: FS, path1: str, fs2: FS, path2: str):
+    from fs.errors import NoSysPath, NoURL
+    from fs.tarfs import ReadTarFS, WriteTarFS
+    from fs.zipfs import ReadZipFS, WriteZipFS
+
+    # def unwrap_wrapfs(fs, path):
+    #     base = "/"
+    #     if isinstance(fs, WrapFS):
+    #         fs, base = fs.delegate_path("/")
+    #     return fs, normpath(join(base, path))  # to support ".." in path
+    # completely unwrap WrapFS instances
+    fs1, path1 = unwrap_wrapfs(fs1, path1)
+    fs2, path2 = unwrap_wrapfs(fs2, path2)
+
+    # obvious check
+    if fs1 == fs2 and path1 == path2:
+        return True
+
+    # check all fs with syspath support
+    try:
+        return fs1.getsyspath(path1) == fs2.getsyspath(path2)
+    except NoSysPath:
+        pass
+
+    # check zip and tar
+    Tar = (WriteTarFS, ReadTarFS)
+    Zip = (WriteZipFS, ReadZipFS)
+    if (isinstance(fs1, Tar) and isinstance(fs2, Tar)) or (
+        isinstance(fs1, Zip) and isinstance(fs2, Zip)
+    ):
+        return path1 == path2 and fs1._file == fs2._file
+
+    # check all fs with url support
+    if isinstance(fs1, fs2.__class__):
+        try:
+            return fs1.geturl(path1) == fs2.geturl(path2)
+        except NoURL:
+            pass
+    return False
+
+
+def safe_description(fs: FS, path):
+    try:
+        if isinstance(fs, SimulationFS):
+            return f"{str(fs)}{fspath.abspath(path)}"
+        return fs.getsyspath(path)
+    except Exception as e:
+        return f'{path} in "{fs}"'
+
+
+def ensure_list(inp):
+    if not isinstance(inp, list):
+        return [inp]
+    return inp
+
+
+def ensure_dict(inp):
+    if isinstance(inp, dict):
+        return inp
+    elif isinstance(inp, str):
+        return {inp: {}}
+    raise ValueError(f"Cannot ensure dict: {inp}")
+
+
+def to_args(inp):
+    """Convert a argument into a (args, kwargs) tuple.
+
+    >>> to_args(None)
+    ([], {})
+    >>> to_args('test')
+    (['test'], {})
+    >>> to_args([1, 2, 3])
+    ([1, 2, 3], {})
+    >>> to_args({'a': {'b': 'c'}})
+    ([], {'a': {'b': 'c'}})
+    >>> to_args([[1, 2, [3, 4], [5, 6]]])
+    ([1, 2, 3, 4, 5, 6], {})
+    """
+    if inp is None:
+        return ([], {})
+    if isinstance(inp, dict):
+        return ([], inp)
+    return (flatten(ensure_list(inp)), {})
 
 
 def flatten(arr: List[Any]) -> List[Any]:
@@ -39,136 +220,36 @@ def flattened_string_list(x, case_sensitive=True) -> Sequence[str]:
     return x
 
 
-def first_key(dic: Mapping) -> Hashable:
-    return list(dic.keys())[0]
-
-
-class DotDict(dict):
+def flatten_all_lists_in_dict(obj):
     """
-    Quick and dirty implementation of a dot-able dict, which allows access and
-    assignment via object properties rather than dict indexing.
-    Keys are case insensitive.
+    >>> flatten_all_lists_in_dict({1: [[2], [3, {5: [5, 6]}]]})
+    {1: [2, 3, {5: [5, 6]}]}
     """
+    if isinstance(obj, dict):
+        for key, value in obj.items():
+            obj[key] = flatten_all_lists_in_dict(value)
+        return obj
+    elif isinstance(obj, list):
+        return [flatten_all_lists_in_dict(x) for x in flatten(obj)]
+    else:
+        return obj
 
-    def __init__(self, *args, **kwargs):
-        super().__init__()
-        # we could just call super(DotDict, self).__init__(*args, **kwargs)
-        # but that won't get us nested dotdict objects
-        od = dict(*args, **kwargs)
-        for key, val in od.items():
-            if isinstance(val, Mapping):
-                value = DotDict(val)
-            else:
-                value = val
-            self[self.normkey(key)] = value
 
-    @staticmethod
-    def normkey(key):
-        if isinstance(key, str):
-            return key.lower()
+def deep_merge(a: dict, b: dict, *, add_keys=True) -> dict:
+    result = deepcopy(a)
+    for bk, bv in b.items():
+        av = result.get(bk)
+        if isinstance(av, dict) and isinstance(bv, dict):
+            result[bk] = deep_merge(av, bv, add_keys=add_keys)
+        elif (av is not None) or add_keys:
+            result[bk] = deepcopy(bv)
+    return result
+
+
+def deep_merge_inplace(base: dict, updates: dict) -> None:
+    for bk, bv in updates.items():
+        av = base.get(bk)
+        if isinstance(av, dict) and isinstance(bv, dict):
+            deep_merge_inplace(av, bv)
         else:
-            return key
-
-    def __delattr__(self, key):
-        try:
-            del self[self.normkey(key)]
-        except KeyError as ex:
-            raise AttributeError("No attribute called: %s" % key) from ex
-
-    def __getattr__(self, key):
-        try:
-            return self[self.normkey(key)]
-        except KeyError as ex:
-            raise AttributeError("No attribute called: %s" % key) from ex
-
-    def __setattr__(self, key, value) -> None:
-        self[self.normkey(key)] = value
-
-    def update(self, other):
-        """ recursively update the dotdict instance with another dicts items """
-        for key, val in other.items():
-            normkey = self.normkey(key)
-            if isinstance(val, Mapping):
-                if isinstance(self.get(normkey), dict):
-                    self[normkey].update(val)
-                else:
-                    self[normkey] = __class__(val)
-            else:
-                self[normkey] = val
-
-    def merge(self, other) -> Mapping:
-        """ recursively merge values from another dict and return a new instance """
-        new_dct = deepcopy(self)
-        new_dct.update(other)
-        return new_dct
-
-
-def increment_filename_version(path: Path, separator=" ") -> Path:
-    stem = path.stem
-    try:
-        # try to find any existing counter
-        splitstem = stem.split(separator)  # raises ValueError on missing sep
-        if len(splitstem) < 2:
-            raise ValueError()
-        counter = int(splitstem[-1])
-        stem = separator.join(splitstem[:-1])
-    except (ValueError, IndexError):
-        # not found, we start with 1
-        counter = 1
-    return path.with_name(
-        "{stem}{sep}{cnt}{suffix}".format(
-            stem=stem, sep=separator, cnt=(counter + 1), suffix=path.suffix
-        )
-    )
-
-
-def find_unused_filename(path: Path, separator=" ") -> Path:
-    """
-    We assume the given path already exists. This function adds a counter to the
-    filename until we find a unused filename.
-    """
-    # TODO: Check whether the assumption can be eliminated for cleaner code.
-    # TODO: Optimization: The counter only needs to be parsed once.
-    tmp = path
-    while True:
-        tmp = increment_filename_version(tmp, separator=separator)
-        if not tmp.exists():
-            return tmp
-
-
-def dict_merge(dct, merge_dct, add_keys=True):
-    """ Recursive dict merge.
-
-    Inspired by :meth:``dict.update()``, instead of
-    updating only top-level keys, dict_merge recurses down into dicts nested
-    to an arbitrary depth, updating keys. The ``merge_dct`` is merged into
-    ``dct``.
-
-    This version will return a copy of the dictionary and leave the original
-    arguments untouched.
-
-    The optional argument ``add_keys``, determines whether keys which are
-    present in ``merge_dict`` but not ``dct`` should be included in the
-    new dict.
-
-    Args:
-        dct (dict) onto which the merge is executed
-        merge_dct (dict): dct merged into dct
-        add_keys (bool): whether to add new keys
-
-    Returns:
-        dict: updated dict
-
-    Taken from comment thread: https://gist.github.com/angstwad/bf22d1822c38a92ec0a9
-    """
-    dct = deepcopy(dct)
-    if not add_keys:
-        merge_dct = {k: merge_dct[k] for k in set(dct).intersection(set(merge_dct))}
-
-    for k, v in merge_dct.items():
-        if isinstance(dct.get(k), dict) and isinstance(v, Mapping):
-            dct[k] = dict_merge(dct[k], v, add_keys=add_keys)
-        else:
-            dct[k] = v
-
-    return dct
+            base[bk] = bv
