@@ -1,150 +1,165 @@
-import os
+from copy import deepcopy
 
-from organize.actions import Copy
-from organize.compat import Path
+import pytest
+from conftest import make_files, read_files
+from fs.base import FS
 
-USER_DIR = os.path.expanduser("~")
+from organize import core
 
-DEFAULT_ARGS = {
-    "basedir": Path.home(),
-    "path": Path.home() / "test.py",
-    "simulate": False,
+files = {
+    "test.txt": "",
+    "file.txt": "Hello world\nAnother line",
+    "another.txt": "",
+    "folder": {
+        "x.txt": "",
+    },
 }
 
 
-def test_tilde_expansion(mock_exists, mock_samefile, mock_copy, mock_trash, mock_mkdir):
-    mock_exists.return_value = False
-    mock_samefile.return_value = False
-    copy = Copy(dest="~/newname.py", overwrite=False)
-    updates = copy.run(**DEFAULT_ARGS)
-    mock_mkdir.assert_called_with(exist_ok=True, parents=True)
-    mock_exists.assert_called_with()
-    mock_trash.assert_not_called()
-    mock_copy.assert_called_with(
-        src=os.path.join(USER_DIR, "test.py"), dst=os.path.join(USER_DIR, "newname.py")
-    )
-    # keep old file path
-    assert updates is None
+@pytest.fixture
+def testfiles(testfs) -> FS:
+    make_files(testfs, files)
+    yield testfs
 
 
-def test_into_folder(mock_exists, mock_samefile, mock_copy, mock_trash, mock_mkdir):
-    mock_exists.return_value = False
-    mock_samefile.return_value = False
-    copy = Copy(dest="~/somefolder/", overwrite=False)
-    copy.run(**DEFAULT_ARGS)
-    mock_mkdir.assert_called_with(exist_ok=True, parents=True)
-    mock_exists.assert_called_with()
-    mock_trash.assert_not_called()
-    mock_copy.assert_called_with(
-        src=os.path.join(USER_DIR, "test.py"),
-        dst=os.path.join(USER_DIR, "somefolder", "test.py"),
-    )
+def test_copy_on_itself(testfiles):
+    config = """
+    rules:
+      - locations: "/"
+        actions:
+          - copy: "/"
+    """
+    core.run(config, simulate=False, working_dir=testfiles)
+    result = read_files(testfiles)
+    assert result == files
 
 
-def test_overwrite(mock_exists, mock_samefile, mock_copy, mock_trash, mock_mkdir):
-    mock_exists.return_value = True
-    mock_samefile.return_value = False
-    copy = Copy(dest="~/somefolder/", overwrite=True)
-    copy.run(**DEFAULT_ARGS)
-    mock_mkdir.assert_called_with(exist_ok=True, parents=True)
-    mock_exists.assert_called_with()
-    mock_trash.assert_called_with(os.path.join(USER_DIR, "somefolder", "test.py"))
-    mock_copy.assert_called_with(
-        src=os.path.join(USER_DIR, "test.py"),
-        dst=os.path.join(USER_DIR, "somefolder", "test.py"),
-    )
-
-
-def test_already_exists(mock_exists, mock_samefile, mock_copy, mock_trash, mock_mkdir):
-    mock_exists.side_effect = [True, False]
-    mock_samefile.return_value = False
-    copy = Copy(dest="~/folder/", overwrite=False)
-    copy.run(**DEFAULT_ARGS)
-    mock_mkdir.assert_called_with(exist_ok=True, parents=True)
-    mock_exists.assert_called_with()
-    mock_trash.assert_not_called()
-    mock_copy.assert_called_with(
-        src=os.path.join(USER_DIR, "test.py"),
-        dst=os.path.join(USER_DIR, "folder", "test 2.py"),
-    )
-
-
-def test_already_exists_multiple(
-    mock_exists, mock_samefile, mock_copy, mock_trash, mock_mkdir
-):
-    mock_exists.side_effect = [True, True, True, False]
-    mock_samefile.return_value = False
-    copy = Copy(dest="~/folder/", overwrite=False)
-    copy.run(**DEFAULT_ARGS)
-    mock_mkdir.assert_called_with(exist_ok=True, parents=True)
-    mock_exists.assert_called_with()
-    mock_trash.assert_not_called()
-    mock_copy.assert_called_with(
-        src=os.path.join(USER_DIR, "test.py"),
-        dst=os.path.join(USER_DIR, "folder", "test 4.py"),
-    )
-
-
-def test_already_exists_multiple_with_separator(
-    mock_exists, mock_samefile, mock_copy, mock_trash, mock_mkdir
-):
-    args = {
-        "basedir": Path.home(),
-        "path": Path.home() / "test_2.py",
-        "simulate": False,
+def test_copy_basic(testfs):
+    config = """
+    rules:
+      - locations: "."
+        filters:
+          - name: test
+        actions:
+          - copy: newname.pdf
+    """
+    make_files(testfs, ["asd.txt", "newname 2.pdf", "newname.pdf", "test.pdf"])
+    core.run(config, simulate=False, working_dir=testfs)
+    assert read_files(testfs) == {
+        "newname.pdf": "",
+        "newname 2.pdf": "",
+        "newname 3.pdf": "",
+        "test.pdf": "",
+        "asd.txt": "",
     }
-    mock_exists.side_effect = [True, True, True, False]
-    mock_samefile.return_value = False
-    copy = Copy(dest="~/folder/", overwrite=False, counter_separator="_")
-    copy.run(**args)
-    mock_mkdir.assert_called_with(exist_ok=True, parents=True)
-    mock_exists.assert_called_with()
-    mock_trash.assert_not_called()
-    mock_copy.assert_called_with(
-        src=os.path.join(USER_DIR, "test_2.py"),
-        dst=os.path.join(USER_DIR, "folder", "test_5.py"),
-    )
 
 
-def test_makedirs(mock_parent, mock_copy, mock_trash):
-    copy = Copy(dest="~/some/new/folder/", overwrite=False)
-    copy.run(**DEFAULT_ARGS)
-    mock_parent.mkdir.assert_called_with(parents=True, exist_ok=True)
-    mock_trash.assert_not_called()
-    mock_copy.assert_called_with(
-        src=os.path.join(USER_DIR, "test.py"),
-        dst=os.path.join(USER_DIR, "some", "new", "folder", "test.py"),
-    )
-
-
-def test_args(mock_exists, mock_samefile, mock_copy, mock_trash, mock_mkdir):
-    args = {
-        "basedir": Path.home(),
-        "path": Path.home() / "test.py",
-        "simulate": False,
-        "nr": {"upper": 1},
+def test_copy_into_dir(testfiles):
+    config = """
+    rules:
+      - locations: "/"
+        actions:
+          - copy: "a brand/new/folder/"
+    """
+    core.run(config, simulate=False, working_dir=testfiles)
+    result = read_files(testfiles)
+    assert result == {
+        "test.txt": "",
+        "file.txt": "Hello world\nAnother line",
+        "another.txt": "",
+        "folder": {
+            "x.txt": "",
+        },
+        "a brand": {
+            "new": {
+                "folder": {
+                    "test.txt": "",
+                    "file.txt": "Hello world\nAnother line",
+                    "another.txt": "",
+                }
+            }
+        },
     }
-    mock_exists.return_value = False
-    mock_samefile.return_value = False
-    copy = Copy(dest="~/{nr.upper}-name.py", overwrite=False)
-    copy.run(**args)
-    mock_mkdir.assert_called_with(exist_ok=True, parents=True)
-    mock_exists.assert_called_with()
-    mock_trash.assert_not_called()
-    mock_copy.assert_called_with(
-        src=os.path.join(USER_DIR, "test.py"), dst=os.path.join(USER_DIR, "1-name.py")
-    )
 
 
-def test_path(mock_exists, mock_samefile, mock_copy, mock_trash, mock_mkdir):
-    mock_exists.return_value = False
-    mock_samefile.return_value = False
-    copy = Copy(dest="~/{path.stem}/{path.suffix}/{path.name}", overwrite=False)
-    copy.run(**DEFAULT_ARGS)
-    mock_mkdir.assert_called_with(exist_ok=True, parents=True)
-    mock_exists.assert_called_with()
-    mock_trash.assert_not_called()
-    mock_copy.assert_called_with(
-        src=os.path.join(USER_DIR, "test.py"),
-        dst=os.path.join(USER_DIR, "test", ".py", "test.py"),
-    )
+def test_copy_into_dir_subfolders(testfiles):
+    config = """
+    rules:
+      - locations: "/"
+        subfolders: true
+        actions:
+          - copy: "a brand/new/folder/"
+    """
+    core.run(config, simulate=False, working_dir=testfiles)
+    result = read_files(testfiles)
+    assert result == {
+        "test.txt": "",
+        "file.txt": "Hello world\nAnother line",
+        "another.txt": "",
+        "folder": {
+            "x.txt": "",
+        },
+        "a brand": {
+            "new": {
+                "folder": {
+                    "test.txt": "",
+                    "file.txt": "Hello world\nAnother line",
+                    "another.txt": "",
+                    "x.txt": "",
+                }
+            }
+        },
+    }
+
+
+@pytest.mark.parametrize(
+    "mode,files,test_txt_content",
+    [
+        ("skip", ["file.txt", "test.txt"], "old"),
+        ("overwrite", ["file.txt", "test.txt"], "new"),
+        ("rename_new", ["file.txt", "test.txt", "test 2.txt"], "old"),
+        ("rename_existing", ["file.txt", "test.txt", "test 2.txt"], "new"),
+    ],
+)
+def test_copy_conflict(testfs: FS, mode, files, test_txt_content):
+    config = f"""
+    rules:
+      - locations: "/"
+        filters:
+          - name: file
+        actions:
+          - copy:
+              dest: "test.txt"
+              on_conflict: {mode}
+    """
+    testfs.writetext("file.txt", "new")
+    testfs.writetext("test.txt", "old")
+    core.run(config, simulate=False, working_dir=testfs)
+    assert set(testfs.listdir("/")) == set(files)
+    assert testfs.readtext("file.txt") == "new"
+    assert testfs.readtext("test.txt") == test_txt_content
+
+
+def test_does_not_create_folder_in_simulation(testfs):
+    config = """
+        rules:
+          - locations: "."
+            actions:
+              - copy: "new-subfolder/"
+              - copy: "copyhere/"
+        """
+    make_files(testfs, files)
+    core.run(config, simulate=True, working_dir=testfs)
+    result = read_files(testfs)
+    assert result == files
+
+    core.run(config, simulate=False, working_dir=testfs)
+    result = read_files(testfs)
+
+    expected = deepcopy(files)
+    expected["new-subfolder"] = deepcopy(files)
+    expected["new-subfolder"].pop("folder")
+    expected["copyhere"] = deepcopy(files)
+    expected["copyhere"].pop("folder")
+
+    assert result == expected
