@@ -1,26 +1,17 @@
 from datetime import datetime, timedelta
 
-from unittest.mock import patch
-
-from organize.compat import Path
 from organize.filters import Created
 
 
 def test_min():
     now = datetime.now()
-    created = Created(days=10, hours=12, mode="older")
-    with patch.object(created, "_created") as mock_cr:
-        mock_cr.return_value = now - timedelta(days=10, hours=0)
-        assert created.run(path=Path("~")) is None
-        mock_cr.return_value = now - timedelta(days=10, hours=13)
-        assert created.run(path=Path("~"))
+    ct = Created(days=10, hours=12, mode="older")
+    assert not ct.matches_datetime(now - timedelta(days=10, hours=0))
+    assert ct.matches_datetime(now - timedelta(days=10, hours=13))
 
 
 def test_max():
     now = datetime.now()
-    created = Created(days=10, hours=12, mode="newer")
-    with patch.object(created, "_created") as mock_cr:
-        mock_cr.return_value = now - timedelta(days=10, hours=0)
-        assert created.run(path=Path("~"))
-        mock_cr.return_value = now - timedelta(days=10, hours=13)
-        assert created.run(path=Path("~")) is None
+    ct = Created(days=10, hours=12, mode="newer")
+    assert ct.matches_datetime(now - timedelta(days=10, hours=0))
+    assert not ct.matches_datetime(now - timedelta(days=10, hours=13))
